got hundred

+++ srcs/prc_wdt_helper.c
@@ -6,7 +6,7 @@
 /*   By: eunskim <eunskim@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/15 18:57:00 by eunskim           #+#    #+#             */
-/*   Updated: 2022/11/17 17:07:33 by eunskim          ###   ########.fr       */
+/*   Updated: 2022/11/17 19:02:53 by eunskim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,7 +59,7 @@ int	get_longer(int prc, int num_len)
 	return (longer);
 }
 
-int	hex_len_check(unsigned long n)
+int	hex_len_check(unsigned int n)
 {
 	int	hex_len;
